remote_agents/pyproject.toml
[project]
name = "AI-Agents"
version = "0.1.1"
description = "AI Agent for CEO using A2A Protocol"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-require-lib",
    "click>=8.1.8",
    "httpx>=0.28.1",
    "langchain-google-genai>=2.0.10",
    "langgraph>=0.3.18",
    "pydantic>=2.10.6",
    "python-dotenv>=1.1.0",
    "langchain-openai",
    "langchain-community",
    "sse-starlette",
    "jwt",
    "jwcrypto",
    "mcp[cli]",
    "openai-agents",
    "langchain-mcp-adapters"
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv.sources]
a2a-require-lib = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
